pyproject.toml
[project]
name = "test-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.57.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.1",
]
